requirements.txt
pytest==7.2.0
requests==2.28.1
allure-pytest==2.9.45
pytest-rerunfailures==10.3
python-dotenv==0.21.1
pyyaml==6.0
pydantic==1.10.4